# perf: Marshal global labels without intermediate map