As global labels aggregation design may stick around for a bit, make sure the performance isn't too bad. When constructing global labels from APMEvent, do not use an intermediate map. Create aggregationpb Global Labels directly.

Benched with a modified Python agent ndjson with metadata `"labels": {"tag0": null, "tag1": "one", "tag2": 2}`

benchstat:
```
goos: linux
goarch: amd64
pkg: github.com/elastic/apm-aggregation/aggregators
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                                             │ bench-out/fix-global-labels-perf-8-before │ bench-out/fix-global-labels-perf-8-after │
                                             │                  sec/op                   │      sec/op        vs base               │
AggregateCombinedMetrics-16                                                  1.873µ ± 4%       1.964µ ±   7%   +4.86% (p=0.009 n=6)
AggregateBatchSerial-16                                                      4.455µ ± 9%       5.195µ ± 123%        ~ (p=0.240 n=6)
AggregateBatchParallel-16                                                    5.062µ ± 6%       5.189µ ±  17%        ~ (p=0.240 n=6)
CombinedMetricsEncoding-16                                                   895.3n ± 2%       914.4n ±   4%   +2.13% (p=0.004 n=6)
CombinedMetricsToBatch-16                                                    63.21µ ± 2%       65.14µ ±   5%   +3.05% (p=0.041 n=6)
EventToCombinedMetrics-16                                                    395.4n ± 3%       394.4n ±   3%        ~ (p=0.699 n=6)
NDJSONSerial/go-2.0.0.ndjson-16                                              15.00m ± 2%       15.02m ±   6%        ~ (p=0.937 n=6)
NDJSONSerial/nodejs-3.29.0.ndjson-16                                         8.867m ± 3%       8.539m ±  12%        ~ (p=0.132 n=6)
NDJSONSerial/python-6.7.2-labels.ndjson-16                                   38.27m ± 5%       32.15m ±   4%  -15.98% (p=0.002 n=6)
NDJSONSerial/python-6.7.2.ndjson-16                                          22.63m ± 3%       22.85m ±   4%        ~ (p=0.180 n=6)
NDJSONSerial/ruby-4.5.0.ndjson-16                                            13.01m ± 3%       13.31m ±   5%        ~ (p=0.485 n=6)
NDJSONParallel/go-2.0.0.ndjson-16                                            14.89m ± 5%       14.92m ±   2%        ~ (p=0.818 n=6)
NDJSONParallel/nodejs-3.29.0.ndjson-16                                       8.944m ± 4%       8.801m ±   2%        ~ (p=0.485 n=6)
NDJSONParallel/python-6.7.2-labels.ndjson-16                                 38.51m ± 2%       32.30m ±   2%  -16.12% (p=0.002 n=6)
NDJSONParallel/python-6.7.2.ndjson-16                                        23.10m ± 2%       22.63m ±   4%        ~ (p=0.240 n=6)
NDJSONParallel/ruby-4.5.0.ndjson-16                                          13.08m ± 3%       13.30m ±   4%        ~ (p=0.394 n=6)
geomean                                                                      682.1µ            678.0µ          -0.61%

                                             │ bench-out/fix-global-labels-perf-8-before │ bench-out/fix-global-labels-perf-8-after │
                                             │                   B/op                    │      B/op        vs base                 │
AggregateCombinedMetrics-16                                               1.105Ki ± 1%        1.102Ki ± 1%        ~ (p=0.206 n=6)
AggregateBatchSerial-16                                                   1.370Ki ± 1%        1.365Ki ± 6%        ~ (p=0.509 n=6)
AggregateBatchParallel-16                                                 1.370Ki ± 2%        1.364Ki ± 1%        ~ (p=0.729 n=6)
CombinedMetricsEncoding-16                                                  0.000 ± 0%          0.000 ± 0%        ~ (p=1.000 n=6) ¹
CombinedMetricsToBatch-16                                                 3.981Ki ± 0%        3.985Ki ± 0%        ~ (p=0.108 n=6)
EventToCombinedMetrics-16                                                   0.000 ± 0%          0.000 ± 0%        ~ (p=1.000 n=6) ¹
NDJSONSerial/go-2.0.0.ndjson-16                                           9.247Mi ± 1%        9.253Mi ± 0%        ~ (p=0.240 n=6)
NDJSONSerial/nodejs-3.29.0.ndjson-16                                      5.552Mi ± 1%        5.575Mi ± 1%        ~ (p=1.000 n=6)
NDJSONSerial/python-6.7.2-labels.ndjson-16                                30.03Mi ± 1%        20.04Mi ± 1%  -33.26% (p=0.002 n=6)
NDJSONSerial/python-6.7.2.ndjson-16                                       15.38Mi ± 0%        15.37Mi ± 1%        ~ (p=0.589 n=6)
NDJSONSerial/ruby-4.5.0.ndjson-16                                         8.248Mi ± 0%        8.248Mi ± 0%        ~ (p=0.699 n=6)
NDJSONParallel/go-2.0.0.ndjson-16                                         9.225Mi ± 0%        9.242Mi ± 1%        ~ (p=0.132 n=6)
NDJSONParallel/nodejs-3.29.0.ndjson-16                                    5.571Mi ± 1%        5.551Mi ± 1%        ~ (p=0.699 n=6)
NDJSONParallel/python-6.7.2-labels.ndjson-16                              29.99Mi ± 1%        19.97Mi ± 1%  -33.41% (p=0.002 n=6)
NDJSONParallel/python-6.7.2.ndjson-16                                     15.35Mi ± 1%        15.35Mi ± 0%        ~ (p=0.937 n=6)
NDJSONParallel/ruby-4.5.0.ndjson-16                                       8.250Mi ± 0%        8.253Mi ± 0%        ~ (p=0.699 n=6)
geomean                                                                                ²                     -4.99%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                             │ bench-out/fix-global-labels-perf-8-before │ bench-out/fix-global-labels-perf-8-after │
                                             │                 allocs/op                 │    allocs/op     vs base                 │
AggregateCombinedMetrics-16                                                 17.00 ± 0%          17.00 ± 0%        ~ (p=1.000 n=6) ¹
AggregateBatchSerial-16                                                     13.00 ± 8%          13.00 ± 8%        ~ (p=1.000 n=6)
AggregateBatchParallel-16                                                   13.00 ± 8%          13.00 ± 0%        ~ (p=1.000 n=6)
CombinedMetricsEncoding-16                                                  0.000 ± 0%          0.000 ± 0%        ~ (p=1.000 n=6) ¹
CombinedMetricsToBatch-16                                                   95.00 ± 0%          95.00 ± 0%        ~ (p=1.000 n=6) ¹
EventToCombinedMetrics-16                                                   0.000 ± 0%          0.000 ± 0%        ~ (p=1.000 n=6) ¹
NDJSONSerial/go-2.0.0.ndjson-16                                            114.0k ± 1%         114.2k ± 1%        ~ (p=0.485 n=6)
NDJSONSerial/nodejs-3.29.0.ndjson-16                                       63.26k ± 2%         63.45k ± 2%        ~ (p=0.485 n=6)
NDJSONSerial/python-6.7.2-labels.ndjson-16                                 326.3k ± 1%         263.9k ± 1%  -19.13% (p=0.002 n=6)
NDJSONSerial/python-6.7.2.ndjson-16                                        162.9k ± 2%         163.0k ± 1%        ~ (p=0.699 n=6)
NDJSONSerial/ruby-4.5.0.ndjson-16                                          98.86k ± 0%         98.89k ± 0%        ~ (p=0.818 n=6)
NDJSONParallel/go-2.0.0.ndjson-16                                          113.9k ± 1%         114.2k ± 0%        ~ (p=0.180 n=6)
NDJSONParallel/nodejs-3.29.0.ndjson-16                                     63.74k ± 1%         63.53k ± 1%        ~ (p=0.729 n=6)
NDJSONParallel/python-6.7.2-labels.ndjson-16                               326.3k ± 1%         264.2k ± 1%  -19.02% (p=0.002 n=6)
NDJSONParallel/python-6.7.2.ndjson-16                                      162.0k ± 1%         162.3k ± 0%        ~ (p=0.589 n=6)
NDJSONParallel/ruby-4.5.0.ndjson-16                                        98.86k ± 0%         98.85k ± 0%        ~ (p=0.732 n=6)
geomean                                                                                ²                     -2.57%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```